fix bbc and itv

+++ plugin.video.itv/service.py
@@ -7,8 +7,6 @@ for root, dirs, files in os.walk(cache):
 
                 
     for f in files:
-        try:os.unlink(os.path.join(root, f))
-        except:pass
-    for d in dirs:
-        try:shutil.rmtree(os.path.join(root, d))
-        except:pass
+        if 'cookies' in f:
+            try:os.unlink(os.path.join(root, f))
+            except:pass
